bench/requirements_bench.txt
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
backcall==0.2.0
beautifulsoup4==4.14.2
bleach==6.1.0
certifi==2025.10.5
charset-normalizer==3.4.3
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
easydict==1.13
einops==0.8.1
et_xmlfile==2.0.0
eva-decord==0.6.1
executing==2.2.1
fairscale==0.4.13
fastjsonschema==2.21.2
filelock==3.16.1
fsspec==2025.3.0
ftfy==6.2.3
hf-xet==1.1.10
huggingface-hub==0.35.3
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.1.4
mpmath==1.3.0
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.1
numpy==1.24.4
omegaconf==2.3.0
openai-clip==1.0.1
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.4
rpds-py==0.20.1
safetensors==0.5.3
scikit-video==1.1.11
scipy==1.10.1
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.3
timm==1.0.20
tinycss2==1.2.1
tokenizers==0.20.3
torch==2.4.1
torchvision==0.19.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
wcwidth==0.2.14
webencodings==0.5.1
wget==3.2
yarg==0.1.9
zipp==3.20.2
